+++ src/app/infuse/InfusionFinder.tsx
@@ -9,7 +9,7 @@ import { DimThunkDispatch } from 'app/store/types';
 import { useEventBusListener } from 'app/utils/hooks';
 import { isD1Item } from 'app/utils/item-utils';
 import clsx from 'clsx';
-import React, { useCallback, useDeferredValue, useEffect, useReducer } from 'react';
+import { useCallback, useDeferredValue, useEffect, useReducer } from 'react';
 import { useSelector } from 'react-redux';
 import { useLocation } from 'react-router';
 import Sheet from '../dim-ui/Sheet';
@@ -260,9 +260,12 @@ export default function InfusionFinder() {
             )}
           </div>
         </div>
-        <div className="infuseSearch">
-          <SearchBar onQueryChanged={onQueryChanged} placeholder={t('Infusion.Filter')} instant />
-        </div>
+        <SearchBar
+          className="infuseSearch"
+          onQueryChanged={onQueryChanged}
+          placeholder={t('Infusion.Filter')}
+          instant
+        />
       </div>
     </div>
   );